Time: 22 ms (41.15%), Space: 54.6 MB (79.69%) - LeetHub

+++ 1921-eliminate-maximum-number-of-monsters/1921-eliminate-maximum-number-of-monsters.java
@@ -1,10 +1,10 @@
 class Solution {
     public int eliminateMaximum(int[] dist, int[] speed) {
-         int n = dist.length;
-        double[] time_to_city = new double[n];
+          int n = dist.length;
+        float[] time_to_city = new float[n];
         
         for (int i = 0; i < n; i++) {
-            time_to_city[i] = (double)dist[i] / speed[i];
+            time_to_city[i] = (float)dist[i] / speed[i];
         }
         
         Arrays.sort(time_to_city);
@@ -16,6 +16,5 @@ class Solution {
         }
         
         return n;
-        
     }
 }
